+++ src/api/IBMi.ts
@@ -3,10 +3,10 @@ import { existsSync } from "fs";
 import * as node_ssh from "node-ssh";
 import os from "os";
 import path, { parse as parsePath } from 'path';
-import { IBMiComponent } from "../components/component";
-import { CopyToImport } from "../components/copyToImport";
-import { CustomQSh } from '../components/cqsh';
-import { ComponentManager } from "../components/manager";
+import { IBMiComponent } from "./components/component";
+import { CopyToImport } from "./components/copyToImport";
+import { CustomQSh } from './components/cqsh';
+import { ComponentManager } from "./components/manager";
 import { CommandData, CommandResult, ConnectionData, IBMiMember, RemoteCommand, WrapResult } from "../typings";
 import { CompileTools } from "./CompileTools";
 import { ConnectionConfiguration } from "./Configuration";
