+++ src/extension.ts
@@ -13,11 +13,11 @@ import * as Debug from './debug';
 import { parseErrors } from "./api/errors/parser";
 import { DeployTools } from "./filesystems/local/deployTools";
 import { Deployment } from "./filesystems/local/deployment";
-import { CopyToImport } from "./components/copyToImport";
-import { CustomQSh } from "./components/cqsh";
-import { GetMemberInfo } from "./components/getMemberInfo";
-import { GetNewLibl } from "./components/getNewLibl";
-import { extensionComponentRegistry } from "./components/manager";
+import { CopyToImport } from "./api/components/copyToImport";
+import { CustomQSh } from "./api/components/cqsh";
+import { GetMemberInfo } from "./api/components/getMemberInfo";
+import { GetNewLibl } from "./api/components/getNewLibl";
+import { extensionComponentRegistry } from "./api/components/manager";
 import { IFSFS } from "./filesystems/ifsFs";
 import { LocalActionCompletionItemProvider } from "./languages/actions/completion";
 import * as Sandbox from "./sandbox";
